[Fix] #94 - Modify KeychainStorage throwable
